Fix issue #11
This commit also drops dependency of all omap motions on side effects (i.e. one-off augroup trick used in https://github.com/svermeulen/vim-NotableFt), because in some vimscript environment, it becomes unreliable and might not be triggered as expected. As a result, the onoremap writing styles for all motions are unified.